Added callback handler for vgicd_changed_istatus in virqmap.c

+++ bmguest/test_vdev_sample.c
@@ -16,7 +16,7 @@ void test_vdev_sample()
     int i;
     int v1, v2, r;
 
-    uart_print( "vdev_sample: Starting test..., base:" ); uart_print_hex32( base ); uart_print("\n\r");
+    uart_print( "vdev_sample: Starting test..., base:" ); uart_print_hex32( (uint32_t) base ); uart_print("\n\r");
     for( i = 0; i < 10; i++ ) {
         v1 = ( 1 + i ) * 2;
         v2 = ( 1 + i ) * 3;
